SNOW-1229108 Add warning about double buffering parameters
