+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/StreamingUtils.java
@@ -1,33 +1,18 @@
 package com.snowflake.kafka.connector.internal.streaming;
 
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.BOOLEAN_VALIDATOR;
 import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.CUSTOM_SNOWFLAKE_CONVERTERS;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG;
 import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.ERRORS_DEAD_LETTER_QUEUE_TOPIC_NAME_CONFIG;
 import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.ERRORS_LOG_ENABLE_CONFIG;
 import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.ERRORS_TOLERANCE_CONFIG;
 import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.ErrorTolerance;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.INGESTION_METHOD_OPT;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.KEY_CONVERTER_CONFIG_FIELD;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_ENABLE_SINGLE_BUFFER;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_MAX_CLIENT_LAG;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.SNOWPIPE_STREAMING_MAX_MEMORY_LIMIT_IN_BYTES;
-import static com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig.VALUE_CONVERTER_CONFIG_FIELD;
 
-import com.google.common.base.Strings;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.Iterables;
-import com.snowflake.kafka.connector.SnowflakeSinkConnectorConfig;
 import com.snowflake.kafka.connector.Utils;
-import com.snowflake.kafka.connector.internal.BufferThreshold;
 import java.time.Duration;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
 import net.snowflake.ingest.streaming.OffsetTokenVerificationFunction;
 import net.snowflake.ingest.utils.Constants;
-import org.apache.kafka.common.config.ConfigException;
 import org.apache.kafka.common.record.DefaultRecord;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -187,162 +172,5 @@ public class StreamingUtils {
   /* Returns dlq topic name if connector config has errors.deadletterqueue.topic.name set */
   public static String getDlqTopicName(Map<String, String> sfConnectorConfig) {
     return sfConnectorConfig.getOrDefault(ERRORS_DEAD_LETTER_QUEUE_TOPIC_NAME_CONFIG, "");
-  }
-
-  /**
-   * Validate Streaming snowpipe related config provided by config(customer's config)
-   *
-   * @param inputConfig given in connector json file
-   * @return map of invalid parameters
-   */
-  public static ImmutableMap<String, String> validateStreamingSnowpipeConfig(
-      final Map<String, String> inputConfig) {
-    Map<String, String> invalidParams = new HashMap<>();
-
-    // For snowpipe_streaming, role should be non empty
-    if (inputConfig.containsKey(INGESTION_METHOD_OPT)) {
-      try {
-        // This throws an exception if config value is invalid.
-        IngestionMethodConfig.VALIDATOR.ensureValid(
-            INGESTION_METHOD_OPT, inputConfig.get(INGESTION_METHOD_OPT));
-        if (inputConfig
-            .get(INGESTION_METHOD_OPT)
-            .equalsIgnoreCase(IngestionMethodConfig.SNOWPIPE_STREAMING.toString())) {
-
-          // check if buffer thresholds are within permissible range
-          invalidParams.putAll(
-              BufferThreshold.validateBufferThreshold(
-                  inputConfig, IngestionMethodConfig.SNOWPIPE_STREAMING));
-
-          invalidParams.putAll(validateConfigConverters(KEY_CONVERTER_CONFIG_FIELD, inputConfig));
-          invalidParams.putAll(validateConfigConverters(VALUE_CONVERTER_CONFIG_FIELD, inputConfig));
-
-          // Validate if snowflake role is present
-          if (!inputConfig.containsKey(Utils.SF_ROLE)
-              || Strings.isNullOrEmpty(inputConfig.get(Utils.SF_ROLE))) {
-            invalidParams.put(
-                Utils.SF_ROLE,
-                Utils.formatString(
-                    "Config:{} should be present if ingestionMethod is:{}",
-                    Utils.SF_ROLE,
-                    inputConfig.get(INGESTION_METHOD_OPT)));
-          }
-
-          /**
-           * Only checking in streaming since we are utilizing the values before we send it to
-           * DLQ/output to log file
-           */
-          if (inputConfig.containsKey(ERRORS_TOLERANCE_CONFIG)) {
-            SnowflakeSinkConnectorConfig.ErrorTolerance.VALIDATOR.ensureValid(
-                ERRORS_TOLERANCE_CONFIG, inputConfig.get(ERRORS_TOLERANCE_CONFIG));
-          }
-          if (inputConfig.containsKey(ERRORS_LOG_ENABLE_CONFIG)) {
-            BOOLEAN_VALIDATOR.ensureValid(
-                ERRORS_LOG_ENABLE_CONFIG, inputConfig.get(ERRORS_LOG_ENABLE_CONFIG));
-          }
-          if (inputConfig.containsKey(ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG)) {
-            BOOLEAN_VALIDATOR.ensureValid(
-                ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG,
-                inputConfig.get(ENABLE_CHANNEL_OFFSET_TOKEN_MIGRATION_CONFIG));
-          }
-
-          if (inputConfig.containsKey(SNOWPIPE_STREAMING_ENABLE_SINGLE_BUFFER)) {
-            BOOLEAN_VALIDATOR.ensureValid(
-                SNOWPIPE_STREAMING_ENABLE_SINGLE_BUFFER,
-                inputConfig.get(SNOWPIPE_STREAMING_ENABLE_SINGLE_BUFFER));
-          }
-
-          if (inputConfig.containsKey(SNOWPIPE_STREAMING_MAX_CLIENT_LAG)) {
-            ensureValidLong(inputConfig, SNOWPIPE_STREAMING_MAX_CLIENT_LAG, invalidParams);
-          }
-
-          if (inputConfig.containsKey(SNOWPIPE_STREAMING_MAX_MEMORY_LIMIT_IN_BYTES)) {
-            ensureValidLong(
-                inputConfig, SNOWPIPE_STREAMING_MAX_MEMORY_LIMIT_IN_BYTES, invalidParams);
-          }
-
-          // Valid schematization for Snowpipe Streaming
-          invalidParams.putAll(validateSchematizationConfig(inputConfig));
-        }
-      } catch (ConfigException exception) {
-        invalidParams.put(
-            INGESTION_METHOD_OPT,
-            Utils.formatString(
-                "Kafka config:{} error:{}", INGESTION_METHOD_OPT, exception.getMessage()));
-      }
-    }
-
-    return ImmutableMap.copyOf(invalidParams);
-  }
-
-  private static void ensureValidLong(
-      Map<String, String> inputConfig, String param, Map<String, String> invalidParams) {
-    try {
-      Long.parseLong(inputConfig.get(param));
-    } catch (NumberFormatException exception) {
-      invalidParams.put(
-          param,
-          Utils.formatString(
-              param + " configuration must be a parsable long. Given configuration" + " was: {}",
-              inputConfig.get(param)));
-    }
-  }
-
-  /**
-   * Validates if key and value converters are allowed values if {@link
-   * IngestionMethodConfig#SNOWPIPE_STREAMING} is used.
-   *
-   * <p>Map if invalid parameters
-   */
-  private static Map<String, String> validateConfigConverters(
-      final String inputConfigConverterField, Map<String, String> inputConfig) {
-    Map<String, String> invalidParams = new HashMap<>();
-
-    if (inputConfig.containsKey(inputConfigConverterField)
-        && DISALLOWED_CONVERTERS_STREAMING.contains(inputConfig.get(inputConfigConverterField))) {
-      invalidParams.put(
-          inputConfigConverterField,
-          Utils.formatString(
-              "Config:{} has provided value:{}. If ingestionMethod is:{}, Snowflake Custom"
-                  + " Converters are not allowed. \n"
-                  + "Invalid Converters:{}",
-              inputConfigConverterField,
-              inputConfig.get(inputConfigConverterField),
-              IngestionMethodConfig.SNOWPIPE_STREAMING,
-              Iterables.toString(DISALLOWED_CONVERTERS_STREAMING)));
-    }
-
-    return invalidParams;
-  }
-
-  /**
-   * Validates if the configs are allowed values when schematization is enabled.
-   *
-   * <p>return a map of invalid params
-   */
-  private static Map<String, String> validateSchematizationConfig(Map<String, String> inputConfig) {
-    Map<String, String> invalidParams = new HashMap<>();
-
-    if (inputConfig.containsKey(SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG)) {
-      BOOLEAN_VALIDATOR.ensureValid(
-          SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG,
-          inputConfig.get(SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG));
-
-      if (Boolean.parseBoolean(
-              inputConfig.get(SnowflakeSinkConnectorConfig.ENABLE_SCHEMATIZATION_CONFIG))
-          && inputConfig.get(VALUE_CONVERTER_CONFIG_FIELD) != null
-          && (inputConfig.get(VALUE_CONVERTER_CONFIG_FIELD).contains(STRING_CONVERTER_KEYWORD)
-              || inputConfig
-                  .get(VALUE_CONVERTER_CONFIG_FIELD)
-                  .contains(BYTE_ARRAY_CONVERTER_KEYWORD))) {
-        invalidParams.put(
-            inputConfig.get(VALUE_CONVERTER_CONFIG_FIELD),
-            Utils.formatString(
-                "The value converter:{} is not supported with schematization.",
-                inputConfig.get(VALUE_CONVERTER_CONFIG_FIELD)));
-      }
-    }
-
-    return invalidParams;
   }
 }
