--- src/test/java/com/snowflake/kafka/connector/SecurityTest.java
@@ -1,48 +1,0 @@
-package com.snowflake.kafka.connector;
-
-import static com.snowflake.kafka.connector.internal.TestUtils.getConfig;
-
-import com.snowflake.kafka.connector.internal.EncryptionUtils;
-import com.snowflake.kafka.connector.internal.FIPSTest;
-import com.snowflake.kafka.connector.internal.TestUtils;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.Map;
-import org.bouncycastle.operator.OperatorCreationException;
-import org.junit.Test;
-
-public class SecurityTest {
-
-  private final IConnectorConfigValidator connectorConfigValidator = new ConnectorConfigValidator();
-
-  @Test
-  public void testRSAPasswordOutput() throws IOException, OperatorCreationException {
-    String testPasswd = "TestPassword1234!";
-    String testKey = FIPSTest.generateAESKey(TestUtils.getPrivateKey(), testPasswd.toCharArray());
-    Map<String, String> testConf = getConfig();
-    testConf.remove(SnowflakeSinkConnectorConfig.SNOWFLAKE_PRIVATE_KEY);
-    testConf.put(SnowflakeSinkConnectorConfig.SNOWFLAKE_PRIVATE_KEY, testKey);
-    testConf.put(SnowflakeSinkConnectorConfig.SNOWFLAKE_PRIVATE_KEY_PASSPHRASE, testPasswd);
-    connectorConfigValidator.validateConfig(testConf);
-    EncryptionUtils.parseEncryptedPrivateKey(testKey, testPasswd);
-    assert !searchInLogFile(testPasswd);
-  }
-
-  static boolean searchInLogFile(String str) throws IOException {
-    String fileName = "sf.log";
-    File log = new File(fileName);
-    FileReader fileReader = new FileReader(log);
-    BufferedReader buffer = new BufferedReader(fileReader);
-    String line;
-    while ((line = buffer.readLine()) != null) {
-      if (line.contains(str)) {
-        return true;
-      }
-    }
-    buffer.close();
-    fileReader.close();
-    return false;
-  }
-}
